requirements.txt
asgiref==3.8.1
Brotli==1.1.0
click==8.1.7
Django==5.1.1
djangorestframework==3.15.2
gunicorn==23.0.0
h11==0.14.0
packaging==24.1
sqlparse==0.5.1
uvicorn==0.31.0
whitenoise==6.7.0
